+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-28T08:16:48.210Z",
+    "2026-05-28T23:42:18.750Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {