+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -409,7 +409,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KSPTH273WQJ2Z2J54QT56DV1"),
+  _id: z.string().default("01KSRFFSH0DBDMA9KJ8D36KD3B"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,